requirements.txt
simple-salesforce
mysql-connector-python
pandas
pyyaml
Faker
